Add NA/EU/TW server IPs

+++ Ekinar/Program.cs
@@ -27,20 +27,28 @@ namespace Ekinar
 
             _handler += new EventHandler(Handler);
             SetConsoleCtrlHandler(_handler, true);
+            Random rand = new Random();
+            string[] prefixes = { "http://127.0.0.1/Game/Vindictus/en-US/EndPoint.txt/", "http://127.0.0.1/EndPoint.txt/" ,"http://127.0.0.1/ko-KR/EndPoint.txt/" };
+            string[] naServerIp = {"34.218.172.146", "34.218.175.106", "34.218.178.68", "34.218.182.12", "34.218.182.202", "34.218.188.180", "34.218.190.181" };
+            string[] euServerIp = {"18.184.128.66", "18.184.151.140", "18.184.195.53", "18.184.196.86", "18.184.197.129", "18.184.198.133" };
+            string[] twServerIp = { "124.108.151.3", "124.108.151.4", "124.108.151.5", "124.108.151.6", "124.108.151.7" };
 
-            string[] prefixes = { "http://127.0.0.1/Game/Vindictus/en-US/EndPoint.txt/", "http://127.0.0.1/EndPoint.txt/" ,
-                                  "http://127.0.0.1./Vindictus/en-EU/EndPoint.txt/",  "http://127.0.0.1/ko-KR/EndPoint.txt/" };
+            
+            int index = rand.Next(naServerIp.Length);
+            // int index = rand.Next(euServerIp.Length);
+            // int index = rand.Next(twServerIp.Length);
 
-            WebServer webServer = new WebServer(WebServerEndPointResponse, "http://127.0.0.1/Game/Vindictus/en-US/EndPoint.txt/");
+            WebServer webServer = new WebServer(WebServerEndPointResponse, prefixes);
             webServer.Run();
-            new EkinarServer().Start(new IPEndPoint(IPAddress.Parse("127.0.0.1"), 27015), new IPEndPoint(IPAddress.Parse("34.218.172.146"), 27015));
+            
+            new EkinarServer().Start(new IPEndPoint(IPAddress.Parse("127.0.0.1"), 27015), new IPEndPoint(IPAddress.Parse(naServerIp[index]), 27015));
             Console.ReadKey();
         }
 
         #region Web Server Response
         public static string WebServerEndPointResponse(HttpListenerRequest request)
         {
-            Console.WriteLine("Received web request from client.");
+            Console.WriteLine("[Ekinar] Received web request from client.");
             string baseDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
             string endpointPath = Path.Combine(baseDirectory, "EndPoint.txt");
             string content = File.ReadAllText(endpointPath);
@@ -89,14 +97,17 @@ namespace Ekinar
             string hostsFilePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.System), @"drivers\etc\hosts");
             string copyHostsFilePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.System), @"drivers\etc\hosts-copy");
 
-            try
+            if (File.Exists(copyHostsFilePath))
             {
-                File.Delete(hostsFilePath);
-                File.Move(copyHostsFilePath, hostsFilePath);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
+                try
+                {
+                    File.Delete(hostsFilePath);
+                    File.Move(copyHostsFilePath, hostsFilePath);
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
             }
         }
         #endregion
